+++ packages/node/src/indexer/dictionary.service.ts
@@ -9,8 +9,8 @@ import {
   gql,
 } from '@apollo/client/core';
 import { Injectable, OnApplicationShutdown } from '@nestjs/common';
-import { buildQuery, GqlNode, GqlQuery, GqlVar, MetaData } from '@subql/common';
 import { DictionaryQueryCondition, DictionaryQueryEntry } from '@subql/types';
+import { buildQuery, GqlNode, GqlQuery, GqlVar, MetaData } from '@subql/utils';
 import fetch from 'node-fetch';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { getLogger } from '../utils/logger';