+++ packages/query/src/yargs.ts
@@ -66,6 +66,11 @@ export function getYargsOption() {
       describe: 'The port the service will bind to',
       type: 'number',
     },
+    'query-complexity': {
+      demandOption: false,
+      describe: 'Level of query complexity',
+      type: 'number',
+    },
     'experimental-historical': {
       demandOption: false,
       describe: 'Add block height argument to queries for fetching historical state of entities',