+++ packages/common-substrate/CHANGELOG.md
@@ -8,6 +8,10 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-05-02
+### Added
+- Add utils package (#928)
+
 ## [0.3.1] - 2022-04-27
 ### Changed
 - Add missing vm2 dependency (#919)