+++ packages/contract-processors/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2022-05-02
+### Changed
+- Update polkadot 8.2.1 (#910)
+
 ## [0.5.4] - 2022-03-22
 ### Added
 - Add support for EIP2930 transactions (#854)